# docs/swagger.yaml
basePath: /api
definitions:
  controller.BookUpdateRequestJson:
    properties:
      data:
        additionalProperties: true
        type: object
    type: object
  feature.BookPayload:
    properties:
      author:
        type: string
      isbn:
        type: string
      title:
        type: string
    type: object
host: localhost:9000
info:
  contact: {}
  description: This is a collection of books API.
  title: Books API
  version: "1.0"
paths:
  /books:
    get:
      description: show list of Book
      parameters:
      - description: Books Title
        in: query
        name: title
        type: string
      - description: Books Author
        in: query
        name: author
        type: string
      - description: Books ISBN
        in: query
        name: isbn
        type: string
      - description: Page. Default is 1
        in: query
        name: page
        type: string
      - description: Limit. Default is 5
        in: query
        name: limit
        type: string
      - description: Order data by field
        in: query
        name: order
        type: string
      - description: Sorting. ASC or DESC
        in: query
        name: sort
        type: string
      responses: {}
      summary: Get Book List
      tags:
      - Book
    post:
      description: Create Book
      parameters:
      - description: body payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/feature.BookPayload'
      responses: {}
      summary: Create Book
      tags:
      - Book
  /books/{id}:
    delete:
      description: Delete Book by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      summary: Delete Book
      tags:
      - Book
    get:
      description: show Book by ID
      parameters:
      - description: book id
        in: path
        name: id
        required: true
        type: string
      responses: {}
      summary: Get one Book
      tags:
      - Book
    put:
      description: 'Update Book with dynamic fields. editable field : isbn, author,
        title'
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: enter desired field that want to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/controller.BookUpdateRequestJson'
      responses: {}
      summary: Update Book
      tags:
      - Book
swagger: "2.0"